+++ CHANGELOG.md
@@ -11,6 +11,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.8.1] - 2021-1-06
+### Changed
+Home page updated  
+
+## [1.8.0] - 2021-1-06
+### Added
+Added preliminary postmortem  
+### Changed
+Home page now links to that ^^^  
+Model now shows actual MOV  
+
 ## [1.7.3] - 2021-1-05
 ### Removed
 Removed banner & live section